build: ui copying to be part of browser's vite build
